Started working on deleteValue
Case 1 and error handling should work fine as they are simple. I think case 2 is working but may need some tweaking. If errors occur in case 2 first check the trailing zeros are being added correctly. Then, check the logic for leftChild as it may need to be changed to resemble the logic for right.